+++ src/bigdata_research_tools/llm/base.py
@@ -86,10 +86,13 @@ class AsyncLLMEngine:
             return AsyncOpenAIProvider(model=self.model)
 
         elif provider == "bedrock":
-            raise NotImplementedError
-            # from bigdata_research_tools.llm.bedrock import BedrockProvider
-            #
-            # return BedrockProvider(model=self.model)
+            from bigdata_research_tools.llm.bedrock import AsyncBedrockProvider
+
+            return AsyncBedrockProvider(model=self.model)
+        elif provider == "azure":
+            from bigdata_research_tools.llm.azure import AsyncAzureProvider
+
+            return AsyncAzureProvider(model=self.model)
         else:
             logger.error(f"Invalid provider: `{self.provider}`")
 
@@ -212,6 +215,10 @@ class LLMEngine:
             from bigdata_research_tools.llm.bedrock import BedrockProvider
 
             return BedrockProvider(model=self.model)
+        elif provider == "azure":
+            from bigdata_research_tools.llm.azure import AzureProvider
+
+            return AzureProvider(model=self.model)
         else:
             logger.error(f"Invalid provider: `{self.provider}`")