+++ src/bigdata_research_tools/tracing.py
@@ -16,7 +16,8 @@ class TraceEventNames(Enum):
     NARRATIVE_MINER = "NarrativeMinersRun"
     THEMATIC_SCREENER = "ThematicScreenerRun"
     RISK_ANALYZER = "RiskAnalyzerRun"
-
+    COMPANY_SEARCH = "CompanySearchRun"
+    RUN_SEARCH = "SearchRun"
 
 @dataclasses.dataclass
 class Trace: