+++ CHANGELOG.md
@@ -18,6 +18,75 @@ Preparation for a first stable release.
 - Renamed `settings` to `utils` to better reflect its purpose as a utility module and free
   the name for future use.
 
+## [0.20.0] - 2025-09-15
+
+### Added
+- Add support for Bigdata.com API keys for authentication. Follow this guide to create your API key: https://docs.bigdata.com/api-rest/introduction
+- Applied the observer pattern to workflows to allow external monitoring of progress during execution. You can find more details on the `bigdata_research_tools.utils.observer` module.
+- Add utilities to create watchlists from a list of company names and to find watchlists by name and optionally fuzzy matching.
+
+
+## [0.19.0] - 2025-09-03
+
+### Added
+- First version of Azure provider included in the LLMEngine
+
+### Changed
+- Now LLM response in the function `generate_theme_tree` uses `repair_json` to clean the response.
+
+## [0.18.1] - 2025-08-28
+
+### Fix
+- Labeler class now choses a parallel call if Bedrock provider is used and concurrent async loop if any other is selected
+- Functions `generate_theme_tree` and `generate_risk_tree` are called passing down the llm_model provided.
+
+### Changed
+- Improved validation of json responses from llms.
+
+## [0.18.0] - 2025-08-25
+
+### Added
+- First version of the async bedrock llm provider
+- Make `__version__` dynamic based on package metadata
+
+### Fix
+- Entity-level motivation was not being returned in the `screen_companies` function
+
+
+## [0.17.3] - 2025-08-08
+
+### Changed
+- Improve logging and tracing
+
+## [0.17.2] - 2025-07-17
+
+### Changed
+- Improve logging and tracing
+
+## [0.17.1] - 2025-07-16
+
+### Changed 
+- Reinstate filter for company entities in screener
+
+## [0.17.0] - 2025-07-16
+
+### Changed 
+- Fix logic to add logo to excel spreadsheets
+- Add post-processing fields to narrative miner
+- Add tracing when calling search_by_companies
+
+## [0.16.0] - 2025-07-10
+
+### Added 
+- Add thematic screener visuals code
+- Add risk scenario workflow, which includes prompts, labeler, visuals, risk theme tree
+
+### Changed 
+- Moved over input validation to query builder
+- Parse Reporting Entity correctly in query builder
+- Show summaries in leaf nodes of theme tree
+- Simplify prompt logic for thematic mindmapper
+
 ## [0.15.1] - 2025-06-12
 
 ### Changed